There are a lot of changes that have been made here, because of the difficulty I had in creating a user (because I didn't try).
continuous.py:
Changed the formatting of all the save files, changed the way that the file is initialized, and changed the name of the directories and files.
Wrote an exception for exiting the code
segmented.py:
much of the same changes, mostly saving names, directories.
Wrote and exception for exiting the code.

run_continuous_script.sh:
runs the contuous script once every set amount of time, which is decded in the python file. All outputs get stored in a log in /home/muonium/Recordinglogs/<YYYY_mmdd>
Goal is to run this with a cron job, which I will talk about in the readme.